backend/requirements.txt
Django==4.0.6
Django==4.0.4
djangorestframework==3.13.1
djangorestframework==3.14.0
python-decouple==3.8
